etl/airflow/dags/requirements.txt
#
# brew install InstantClientTap/instantclient/instantclient-sqlplus
# python3 -mpip install -r requirements.txt
#

pandas
fastparquet
pyspark==3.4.0
boto3
pyyaml
pymongo
kafka-python